requirements.txt
aws-cdk-lib==2.148.0
aws-cdk.lambda-layer-kubectl-v30==2.0.0
cdk-eks-karpenter==1.0.6
constructs>=10.0.0,<11.0.0
